Added backward compatible raster property on HeatMap
