+++ holoviews/plotting/bokeh/raster.py
@@ -9,8 +9,9 @@ except ImportError:
 
 from ...core.util import cartesian_product, is_nan, unique_array
 from ...element import Image, Raster, RGB
+from ...element.util importget_2d_aggregate
 from ..renderer import SkipRendering
-from ..util import map_colors, get_2d_aggregate
+from ..util import map_colors
 from .element import ElementPlot, ColorbarPlot, line_properties, fill_properties
 from .util import mplcmap_to_palette, get_cmap, hsv_to_rgb, mpl_to_bokeh
